2_rag_pdf/requirements.txt
streamlit==1.32.0
chromadb==0.4.22
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.0
